+++ commands/rules.js
@@ -1,0 +1,4 @@
+
+exports.module = async (message) => { 
+    return `Aturan dibuat buat di langgar, makin sering *toxic* *reputasi* lu *ancur*\ncek reputasi !stat.\ngaboleh ngirim link *bokep* di sini kalo mau japri,\nyg mau transaksi silahkan di japri juga\nOke???\n\n*W0lv*`
+}
